+++ src/py42/_internal/clients/detection_list_user.py
@@ -1,6 +1,5 @@
 import json
 
-from py42._internal.compat import str
 from py42.clients import BaseClient
 from py42.exceptions import Py42BadRequestError, Py42NotFoundError
