+++ src/py42/sdk/queries/fileevents/file_event_query.py
@@ -1,5 +1,3 @@
-import json
-
 from py42._internal.compat import str
 from py42.sdk.queries import BaseQuery
 from py42.sdk.queries.query_filter import (
